+++ simple_history/utils.py
@@ -2,6 +2,7 @@ import warnings
 
 import django
 from django.db import transaction
+from django.db.models import ManyToManyField
 from django.forms.models import model_to_dict
 
 from simple_history.exceptions import NotHistoricalModelError
@@ -69,7 +70,13 @@ def bulk_create_with_history(
         record
     :return: List of objs with IDs
     """
-
+    # Exclude ManyToManyFields because they end up as invalid kwargs to
+    # model.objects.filter(...) below.
+    exclude_fields = [
+        field.name
+        for field in model._meta.get_fields()
+        if isinstance(field, ManyToManyField)
+    ]
     history_manager = get_history_manager_for_model(model)
     second_transaction_required = True
     with transaction.atomic(savepoint=False):
@@ -88,7 +95,10 @@ def bulk_create_with_history(
         with transaction.atomic(savepoint=False):
             for obj in objs_with_id:
                 attributes = dict(
-                    filter(lambda x: x[1] is not None, model_to_dict(obj).items())
+                    filter(
+                        lambda x: x[1] is not None,
+                        model_to_dict(obj, exclude=exclude_fields).items(),
+                    )
                 )
                 obj_list += model.objects.filter(**attributes)
             history_manager.bulk_history_create(